requirements.txt
alembic==1.4.3
appdirs==1.4.4
asgiref==3.2.10
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.1
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.5
cssselect2==0.4.1
cycler==0.11.0
distlib==0.3.1
Django==3.1.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
et-xmlfile==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fonttools==4.27.1
gunicorn==19.9.0
html5lib==1.1
idna==2.10
importlib-metadata==2.1.3
importlib-resources==5.4.0
isoweek==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
MouseInfo==0.1.3
openpyxl==3.1.2
paho-mqtt==1.6.1
pdfkit==1.0.0
pipenv==2018.11.26
PyAutoGUI==0.9.50
pycparser==2.20
pydyf==0.1.2
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyparsing==3.1.1
pyperclip==1.8.0
pyphen==0.11.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.20.0
python-editor==1.0.4
PyTweening==1.0.3
pytz==2020.1
pywhatkit==2.7
requests==2.24.0
russian-names==0.1.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
svgwrite==1.4
tinycss2==1.1.1
Tree==0.2.4
tzdata==2023.3
urllib3==1.25.10
virtualenv==20.0.33
virtualenv-clone==0.5.4
visitor==0.1.3
weasyprint==54.3
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
wkhtmltopdf-wrapper==0.3.3
WTForms==2.3.3
zipp==3.6.0
zopfli==0.1.9
psycopg2==2.9.3